% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/park_palette.R
\name{list_park_palettes}
\alias{list_park_palettes}
\title{A function to list available palettes.}
\usage{
list_park_palettes(pal = park.pal)
}
\description{
This function returns a list of the available palettes in the package.
}
\examples{
list_park_palettes()
}
