% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/park_palette.R
\name{park_palette}
\alias{park_palette}
\title{A function to return the palette based on user inputs.}
\usage{
park_palette(name, n = 10)
}
\arguments{
\item{name}{Name of national park palette (call \code{list_park_palettes()} for available palettes).}

\item{n}{Number of colors from palette. Default is the maximum of 10.}
}
\description{
This function returns the palette based on the desired name and size.
}
\examples{
park_palette()
}
